dfacto_backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0
python-dotenv>=1.0.0
pydantic>=2.7.0
google-genai>=1.0.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-tavily>=0.2.0
tavily-python>=0.3.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
